refactor(mobile): split fizruk dualWrite/diff.ts into per-shape module folder (P2.2a)
Decompose apps/mobile/src/modules/fizruk/lib/dualWrite/diff.ts (633 LOC)
into a module-folder mirroring the operation-family layout of
dualWrite/adapter.ts. Each per-shape file owns its own op types,
snapshot interface, and diff<Shape>Ops function:

  - workouts.ts (95 LOC)        workout-upsert / workout-delete
  - customExercises.ts (45)     custom-exercise-upsert / -delete
  - measurements.ts (46)        measurement-upsert / -delete
  - dailyLog.ts (66)            daily-log-upsert / -delete
  - monthlyPlan.ts (41)         monthly-plan-set
  - workoutTemplates.ts (65)    workout-template-upsert / -delete
  - programs.ts (39)            programs-set
  - planTemplate.ts (35)        plan-template-set
  - wellbeing.ts (80)           wellbeing-upsert / -delete
  - activeWorkout.ts (38)       active-workout-set
  - diffArray.ts (44)           shared per-id diff helper
  - index.ts (229)              barrel: orchestrator + state + union
                                + public re-exports

The barrel preserves the historical public API byte-for-byte
(`from "./dualWrite/diff"` continues to resolve via folder
+ index.ts); no consumer was touched (fizrukDualWriteState.ts,
residualImport.ts, fizrukBackup.ts, __tests__/diff.test.ts).
Largest new file is 229 LOC, well under the Hard Rule #18 600
threshold.

Tests:
  - existing __tests__/diff.test.ts (24 cases) passes unchanged
  - added __tests__/diff.perShape.test.ts (21 cases — 1 happy-path
    + 1 edge-case for each of the 10 shapes + the diffArray
    sort/equality contract)

Audit: docs/audits/2026-05-13-mobile-reliability-ux-roast.md § P2.2a
       (marked closed; outstanding row 3/12 -> 4/12)
