+++ docs/audits/README.md
@@ -44,7 +44,7 @@
 | [`2026-05-13-backend-performance-roast.md`](./2026-05-13-backend-performance-roast.md)             | Backend/performance прожарка #4/10 — error handling, validation, observability, env                                            | Active | 7/14 ≈           | 7 ≈           | embedded "Прогрес виконання" + cross-refs                                                                    |
 | [`2026-05-13-documentation-hygiene-roast.md`](./2026-05-13-documentation-hygiene-roast.md)         | Doc-hygiene прожарка #8/10 — links/freshness/playbook-catalog, post-archive-move regression fix                                | Active | 4/4 P0           | 3 P1 / 2 P2   | embedded §Progress                                                                                           |
 | [`2026-05-13-ftux-onboarding-roast.md`](./2026-05-13-ftux-onboarding-roast.md)                     | FTUX onboarding day 0-7 roast (прожарка #1/10)                                                                                 | Active | 3/3 P0 + 1/4 P1  | 3 P1 + 3 P2   | [`docs/launch/product-os/ftux-master-tracker.md`](../launch/product-os/ftux-master-tracker.md)               |
-| [`2026-05-13-mobile-reliability-ux-roast.md`](./2026-05-13-mobile-reliability-ux-roast.md)         | Прожарка #10/10 — mobile (Expo + Capacitor) reliability & UX                                                                   | Active | 3/12             | 9             | embedded "Outstanding" table                                                                                 |
+| [`2026-05-13-mobile-reliability-ux-roast.md`](./2026-05-13-mobile-reliability-ux-roast.md)         | Прожарка #10/10 — mobile (Expo + Capacitor) reliability & UX                                                                   | Active | 4/12             | 8             | embedded "Outstanding" table                                                                                 |
 | [`2026-05-13-security-observability-roast.md`](./2026-05-13-security-observability-roast.md)       | Security & observability roast — CSP, secrets, Sentry/OTel/web-vitals, audit logs                                              | Active | 1/9              | 8             | embedded §Прогрес-виконання + §P0/P1/P2                                                                      |
 | [`2026-05-13-testing-devx-roast.md`](./2026-05-13-testing-devx-roast.md)                           | Прожарка #6/10 — Testing & DevX (Vitest/Jest/Playwright/Detox, CI, pre-commit)                                                 | Active | 8/14             | 6             | embedded §Outstanding + [`docs/testing/2026-05-05-tests-pr-plan.md`](../testing/2026-05-05-tests-pr-plan.md) |
 | [`2026-05-13-web-architecture-state-roast.md`](./2026-05-13-web-architecture-state-roast.md)       | Roast #3/10 — Web Architecture & State (provider invariant test, typed standalone routes registry, state-write-paths doctrine) | Active | 5/5 closed in PR | 5 P1 deferred | self (§1.1 / §1.2 / §2.1 closed; P1-A/B/C/D/E carried forward)                                               |
